% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{con_postgres}
\alias{con_postgres}
\title{Create a connection to Postgres}
\usage{
con_postgres(configuration_header = "postgres_docker")
}
\arguments{
\item{configuration_header}{A header within the configuration file}

\item{configuration_location}{Pathway to the configuration file}
}
\value{
A database connection to postgres
}
\description{
docker pull postgres
mkdir -p $HOME/docker/volumes/postgres
docker run --rm  --name pg-docker -e POSTGRES_PASSWORD=docker -d -p 5432:5432 -v $HOME/docker/volumes/postgres:/var/lib/postgresql/data  postgres
psql -h localhost -U postgres -d postgres
}
\examples{

\dontrun{
con <-
  postgres_connect(configuration_header = "postgres", configuration_location = 'config.ini')
}
}
