% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_add.R
\name{config_add}
\alias{config_add}
\title{Displays current configuration file}
\usage{
config_add(header = NULL, ...)
}
\arguments{
\item{header}{The refrence to the config parameters within config.ini}

\item{...}{dplyr like syntax for creating ini}
}
\description{
Displays current configuration file
}
