% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_default.R
\docType{data}
\name{config_default}
\alias{config_default}
\title{Default config file}
\format{An object of class \code{character} of length 1.}
\usage{
config_default
}
\description{
Default config file
}
\keyword{datasets}
