% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_remove.R
\name{config_remove}
\alias{config_remove}
\title{Removes value from ini}
\usage{
config_remove(header = NULL)
}
\arguments{
\item{header}{The refrence to the config parameters within config.ini}
}
\description{
Removes value from ini
}
