% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_loop_to.R
\name{copy_loop_to}
\alias{copy_loop_to}
\title{copy_loop_to transfers table from one database to another}
\usage{
copy_loop_to(conn1, conn2 = NULL, name, n = NULL, statement)
}
\arguments{
\item{conn1:}{Connection to initial database where the table is coming from.}

\item{conn2:}{Connection to the database where the table is transferring to. If conn2 is NULL then conn2 will default to initial database.}

\item{name:}{Name of table on conn2 database}

\item{n:}{Batch loader, amount of data to be transferred with each cycle.}

\item{statement:}{SQL statement on initial database.}
}
\value{
Transfers table from intital database to another.
}
\description{
Transfers table from one database to the another without overloading R memory. Best used when databases don't have a connection to eachother.
}
\examples{
 #Sends table (created from sql statement) from post database to tera database with a batch load of 100.
   copy_loop_to(post, tera, "testdata", 100, "SELECT * FROM TESTTABLE")
}
