% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_indexes.R
\name{create_indexes}
\alias{create_indexes}
\title{create_indexes creates indexes onto an existing table}
\usage{
create_indexes(con, table, indexes = NULL, unique = FALSE, ...)
}
\arguments{
\item{conn:}{Connection to database. Can used assigned output from function connectR for the connection.}

\item{table:}{Name of table on the database}

\item{indexes:}{a vector of column names to become indexes for a table.}

\item{unique:}{Specifies if the indexes are unique or not (TRUE or FALSE).}
}
\value{
Indexes on a table
}
\description{
Creates indexes onto an existing table within a database
}
\examples{
 #Creates a table with no data on the database, but allows primary
 keys/indexes to be created:
   create_indexes(post, "testdata", testdata, c("col1","col2"))
}
