% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_column_info.R
\name{db_column_info}
\alias{db_column_info}
\title{db_column_info returns column information from a SQL query}
\usage{
db_column_info(conn, statement)
}
\arguments{
\item{conn:}{Connection to database. Can used assigned output from function connectR for the connection.}

\item{statement:}{SQL query}
}
\value{
A Tibble containing column information from a SQL query
}
\description{
Creates a tibble with information about the columns from a SQL query
}
\examples{
 #Creates a tibble/ data.frame with information about
  the columns in a SQL query
   db_column_info(tera,"select * from cis.cmi_daily_demand")
}
