% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create_primary.R
\name{db_create_primary}
\alias{db_create_primary}
\title{db_create_primary creates empty table with primary indexes/ keys}
\usage{
db_create_primary(conn, name = NULL, value = NULL, primary = NULL,
  query = FALSE)
}
\arguments{
\item{conn:}{Connection to database. Can used assigned output from function connectR for the connection.}

\item{name:}{Name of table on the database}

\item{value:}{Data.frame}

\item{primary:}{Column(s) to created to primary keys / indexes.}
}
\value{
Empty table in database with primary keys/ indexes
}
\description{
Creates a table in the database from a dataframe and creates primary keys/ indexes
}
\examples{
 #Creates a table with no data on the database, but allows primary
 keys/indexes to be created:
   db_create_primary(post, "testdata", testdata, c("col1","col2"))
}
