% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_disconnect.R
\name{db_disconnect}
\alias{db_disconnect}
\title{Disconnect from a connection to Postgres}
\usage{
db_disconnect(connection = NULL)
}
\arguments{
\item{connection}{An active connection}
}
\description{
Disconnect from a connection to Postgres
}
\examples{
\dontrun{
con <-
  postgres_connect(configuration_header = "postgres", configuration_location = 'config.ini')

 db_disconnect(connection = con)
}
}
