% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_insert.R
\name{db_insert}
\alias{db_insert}
\title{db_insert inserts data into a database table from a data.frame}
\usage{
db_insert(con, name, values, ...)
}
\arguments{
\item{conn:}{Connection to database. Can used assigned output from function connectR for the connection.}

\item{name:}{Table name on database data is inserting into.}

\item{values:}{Data.frame that is being sent to database.}
}
\value{
Inserts data.
}
\description{
Inserts data into a database table from a data.frame
}
\examples{
 #Inserts testtable data into the table "testtable" on the database.
   db_insert(post,"testtabe",testtable)
}
