% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_rename_column.R
\name{db_rename_column}
\alias{db_rename_column}
\title{db_rename_column_name will reformat column names align to databases}
\usage{
db_rename_column(con, value, col_name = "Col")
}
\arguments{
\item{conn:}{Connection to database. Can used assigned output from function connectR for the connection.}

\item{value:}{Data.frame}

\item{col_name:}{Column name format for columns that start with a number}
}
\value{
A vector with reformatted columns names from input data.frame
}
\description{
Creates a vector with reformatted columns names from input data.frame
}
\examples{
 #Creates a vector with reformatted column names and any column begining with a
   number with renamed to `Col[0-9]`:
   db_rename_column(post, testdata)
}
