% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbtbls.R
\name{dbtbls}
\alias{dbtbls}
\title{dbtbls function}
\usage{
dbtbls(conn, db = NULL, schema = NULL)
}
\arguments{
\item{conn:}{Connection to database. Can used assigned output from function connectR for the connection.}

\item{db:}{Database name i.e. default}

\item{schema:}{Database schema i.e public}
}
\value{
Connection to database
}
\description{
Function searches tables in entire datadata / schema.
}
\examples{
 #Search for default user (account signed into the computer)
  tables in the database default.
   dbtbls(post)->dbtables

 #Search for default user (account signed into the computer)
  tables in the database c("postgres","postgres1"), also searches for schema public.
   dbtbls(post,db=c("postgres","postgres1"),schema="public")->dbtables
}
