% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_postgres.R
\docType{data}
\name{docker_postgres}
\alias{docker_postgres}
\title{Dockerfile for postgres}
\format{An object of class \code{character} of length 1.}
\usage{
docker_postgres
}
\description{
Dockerfile for postgres
}
\keyword{datasets}
