% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_files.R
\name{generate_files}
\alias{generate_files}
\title{Creates default files for connectr}
\usage{
generate_files(connectr_base_dir = "~")
}
\arguments{
\item{connectr_base_dir}{Where the connectr file should be located}
}
\description{
Creates default files for connectr
}
