% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{postgres_connection}
\alias{postgres_connection}
\title{Create a connection to Postgres}
\usage{
postgres_connection(configuration_header = "postgres",
  configuration_location = "~/.config.ini")
}
\arguments{
\item{configuration_header}{A header within the configuration file}

\item{configuration_location}{Pathway to the configuration file}
}
\value{
A database connection to postgres
}
\description{
Create a connection to Postgres
}
\examples{

\dontrun{
con <-
  postgres_connect(configuration_header = "postgres", configuration_location = 'config.ini')
}
}
