% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_query.R
\name{save_query}
\alias{save_query}
\title{save_query saves connectR's sql to string for tbl, and optional saves to local file.}
\usage{
save_query(x, save_sql = FALSE, path = NULL, name = NULL, type = NULL)
}
\arguments{
\item{x:}{Lazy evaluation object. Object created from the function `tbl`.}

\item{save_sql:}{Saves connectR sql when set to `TRUE`, default set to `FALSE`.}

\item{path:}{Path to save file, default set to work directory `getwd()`.}

\item{name:}{Name of the file, default set to "connectR_sql". Name is only needed and file type is not needed.}

\item{type:}{File type for saving local file for example `.txt`, defaults to `.SQL`.}
}
\value{
Returns string format of connectR's sql, and optional save to local file.
}
\description{
Allows to save connectR's sql query in local file and save to a string for further tbl's. NOTE: save_query will only work with lazy evaluation objects.
}
\examples{
#Save lazy evaluation sql to a string
   tbl(post,sql("SELECT * FROM
                 CIS.CMI_CALENDAR")) \%>\%
   group_by(ACCOUNT_YEAR_WEEK,
            SEASON) \%>\%
   summarise(WEEK_ENDING=max(CAL_DATE)) \%>\%
   save_query ->SQL
#Uses sql string for further lazy_evaluation querying
   tbl(post,sql(SQL))

#Save lazy evaluation sql to default location `getwd()`
#with default name `connectR_sql`
   tbl(post,sql("SELECT * FROM
                 CIS.CMI_CALENDAR")) \%>\%
   group_by(ACCOUNT_YEAR_WEEK,
            SEASON) \%>\%
   summarise(WEEK_ENDING=max(CAL_DATE)) \%>\%
   save_query(save_query=TRUE)

#Save lazy evaluation sql to default location `C:/R_CODE`
#with default name `connectR_sql`
   tbl(post,sql("SELECT * FROM
                 CIS.CMI_CALENDAR")) \%>\%
   group_by(ACCOUNT_YEAR_WEEK,
            SEASON) \%>\%
   summarise(WEEK_ENDING=max(CAL_DATE)) \%>\%
   save_query(save_query=TRUE, path="C:/R_CODE")

#Save lazy evaluation sql to default location `C:/R_CODE`
#with name `Test_sql`
   tbl(post,sql("SELECT * FROM
                 CIS.CMI_CALENDAR")) \%>\%
   group_by(ACCOUNT_YEAR_WEEK,
            SEASON) \%>\%
   summarise(WEEK_ENDING=max(CAL_DATE)) \%>\%
   save_query(save_query=TRUE, path="C:/R_CODE", name="Test_sql")

#Save lazy evaluation sql to default location `C:/R_CODE`
#with name `Test_sql` and type ".txt" text file.
   tbl(post,sql("SELECT * FROM
                 CIS.CMI_CALENDAR")) \%>\%
   group_by(ACCOUNT_YEAR_WEEK,
            SEASON) \%>\%
   summarise(WEEK_ENDING=max(CAL_DATE)) \%>\%
   save_query(save_query=TRUE, path="C:/R_CODE", name="Test_sql",type=".txt")
}
