% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-sample.R
\name{top_df}
\alias{top_df}
\title{Select top n rows returns data.frame}
\usage{
top_df(x = NULL, n = NULL)
}
\arguments{
\item{x:}{Lazy evaluation object. Object created from the function `tbl`.}

\item{n:}{Number of rows, n defaults to 10.}
}
\value{
Returns a data.frame with n rows from a lazy evaluation object.
}
\description{
top_df gets the top n of rows from lazy evaluation object and returns a data.frame.
}
\examples{
#Top 20 rows from lazy evaluation object
   tbl(post,sql("SELECT * FROM
                 CALENDAR")) \%>\%
   group_by(ACCOUNT_YEAR_WEEK,
            SEASON) \%>\%
   summarise(WEEK_ENDING=max(CAL_DATE)) \%>\%
   top_df(n=20)
}
