% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectR.R
\name{ConnectR}
\alias{ConnectR}
\title{Title}
\usage{
ConnectR(
  file,
  type,
  sheets = "",
  indices = "",
  n_boot = 1000,
  dist_type = "density"
)
}
\arguments{
\item{file}{Name of the file}

\item{type}{Type of the file (.xlsx or .csv)}

\item{sheets}{Number of sheets to process}

\item{indices}{Indices to calculate}

\item{n_boot}{Number on bootstraps samples for resampling}

\item{dist_type}{distibution type: "histogram" or "density"}
}
\value{
an assessment of the biogeographic structuring of the bipartite network provided
}
\description{
Title
}
\examples{
ConnectR("Aulopiformesdata.xlsx",".xlsx")
}
