% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectR.R
\name{connectr}
\alias{connectr}
\title{ConnectR}
\usage{
connectr(
  file,
  type,
  sheets = "",
  indices = "",
  n_boot = 1000,
  dist_type = "density",
  color = c("blue", "red")
)
}
\arguments{
\item{file}{Rute to the file to analize}

\item{type}{Type of the file either "excel" or "csv"}

\item{sheets}{Quantity of sheets (if there's any) to evaluate}

\item{indices}{Quantity of indices to evaluate.}

\item{n_boot}{Quantity of samples to perform a parametric bootstrap (1000 set by default)}

\item{dist_type}{How to display the bootstrap (density set by default)}

\item{color}{Range of the palette of colors to display the graphics (scale from blue to red by default)}
}
\value{
Dashboard full of information from the indices and the distribution
graphs for the bootstraps sets.
}
\description{
This package takes a file with the correct format, which can be "xlsx" or "csv".
And perform an parametric bootstrap re-sampling the data for the four coefficients (by now)
such as Biogeographical Connectiveness \link{"BC"}, Average Occurrences \link{"AO"}, Average Endemics \link{"AE"}
and Network Clustering \link{"NC"}. This also displays a Dashboard with the values of each coefficient
along side a timer which shows theremain time execution operation. At last it will generate an image
with the arrengment of the grahps for every distribution obtained with the parametric bootstrap method.
}
\examples{
connectr(file= "Aulopiformes.xlsx",type= "excel",sheets= c(2)
,indices=c("BC","AE","NC"),n_boot=3000, color = c("green", "yellow"))

}
