% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_send_query.R
\name{db_send_query}
\alias{db_send_query}
\title{db_send_query sends SQL}
\usage{
db_send_query(conn, statement, ...)
}
\arguments{
\item{con:}{Connection to database. Can used assigned output from function connectR for the connection.}

\item{statement:}{Sql query to be sent to the database.}

\item{...:}{Other parameters passed on to methods.}
}
\value{
Sends sql queries directly to the database.
}
\description{
Sends sql queries to the database.
}
\examples{
 #Sending sql query to create table on database.
   db_send_query(post, "create table hello as (select * from world) with no data;")
}
