% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbtbls.R
\name{dbtbls}
\alias{dbtbls}
\title{dbtbls function}
\usage{
dbtbls(conn, uid = NULL, sbstr = NULL, db = NULL, schema = NULL)
}
\arguments{
\item{conn:}{Connection to database. Can used assigned output from function connectR for the connection.}

\item{uid:}{User id for database}

\item{sbstr:}{A 3 character string, that searches table names i.e. "DL_". Can set sbstr to be a vector for multiple search i.e. c("DL_","DJ_")}

\item{db:}{Database name i.e. TMG}

\item{schema:}{Database schema i.e public}
}
\value{
Connection to database
}
\description{
Function that connects to data bases, and also ecrypt password for safe storage.
}
\examples{
 #Search for default user (account signed into the computer)
  tables in the database TMG.
   dbtbls(zeus)->dbtables

 #Search for default user (account signed into the computer)
  tables in the database TMG, also searches for c("DJ_","DRJ") and
   at start of table name variable.
   dbtbls(zeus,sbstr=c("DJ_","DRJ"))->dbtables

 #Search for default user, (account signed into the computer)
  tables in the database postgres, shopdirect and also searches in the schema public
   at start of table name variable.
   dbtbls(post,sbstr=c("DJ_","DRJ"), db=c("postgres","shopdirect"), schema="public")->dbtables
}
