% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectR.R
\name{src_connectR}
\alias{src_connectR}
\title{src_connectR connect to Databases}
\usage{
src_connectR(dsn = NULL, uid = NULL, pwd = NULL, Update = FALSE,
  db = NULL, ..., auto_disconnect = FALSE)
}
\arguments{
\item{dsn:The}{Data Source Name.}

\item{uid:}{User id for database connecting to.}

\item{pwd:}{User password for database}

\item{db:}{Database name i.e. TMG}

\item{Update:}{Update encrypted password}

\item{...:}{Additional ODBC keywords, these will be joined with the other arguments to form the final connection string.}
}
\value{
Connection to database
}
\description{
Connects to databases currently using the odbc api, and also ecrypt password for safe storage.
}
\examples{
 #Standard connection to teradata and to database "postgres",
 unadviced as password in contained within the code:
   src_connectR(dsn = "POSTGRES", uid = "USERNAME", pwd = "PaSsword",
    db = "postgres", Update = F)

 #Standard connection to teradata and to all database,
 unadviced as password in contained within the code:
   src_connectR(dsn = "POSTGRES", uid = "USERNAME", pwd = "PaSsword",
    Update = F)

 #Connection to teradata and to all databases,
 R will take store encrypted username and password:
   connectR(dsn = "POSTGRES")

 #Connection to teradata and to all databases,
 as Update = TRUE, stored password will be updated:
   connectR(dsn = "POSTGRES",Update=T)
}
